\name{ConfirmDnaSeq}
\alias{ConfirmDnaSeq}
\title{Confirms character is a valid DNA sequence.}
\usage{
  ConfirmDnaSeq(dna.seq)
}
\description{
  This function checks that a DNA sequence in character
  string format, e.g. "GATTACA", and confirms it consists
  only of upper or lower-case character vector consisting
  of A, C, T, G, or any of the IUPAC ambiguous nucleotide
  letters.
}
\keyword{vectorizer}

