\name{KmerGenerator}
\alias{KmerGenerator}
\title{Creates a vector of kmers for a given k value.}
\usage{
KmerGenerator(kmer)
}
\arguments{
  \item{kmer}{an integer that defines the kmer length to be
  calculated}
}
\description{
This function creates a vector of length 4^k consisting of
the kmers for a given value of k.
}
\keyword{vectorizer}

