\name{UpperCaser}
\alias{UpperCaser}
\title{Converts a character string to upper-case letters.}
\usage{
UpperCaser(dna.seq)
}
\arguments{
  \item{dna.seq}{an upper or lower-case character vector
  consisting of A, C, T, G, or any of the IUPAC ambiguous
  nucleotide letters}
}
\description{
This function converts a character string of letters to an
upper-case character string of letters.
}
\keyword{vectorizer}

